+++ moor/test/data/tables/custom_tables.g.dart
@@ -10,16 +10,20 @@ part of 'custom_tables.dart';
 class Config extends DataClass implements Insertable<Config> {
   final String configKey;
   final String configValue;
-  Config({@required this.configKey, this.configValue});
+  final SyncType syncState;
+  Config({@required this.configKey, this.configValue, this.syncState});
   factory Config.fromData(Map<String, dynamic> data, GeneratedDatabase db,
       {String prefix}) {
     final effectivePrefix = prefix ?? '';
     final stringType = db.typeSystem.forDartType<String>();
+    final intType = db.typeSystem.forDartType<int>();
     return Config(
       configKey: stringType
           .mapFromDatabaseResponse(data['${effectivePrefix}config_key']),
       configValue: stringType
           .mapFromDatabaseResponse(data['${effectivePrefix}config_value']),
+      syncState: ConfigTable.$converter0.mapToDart(intType
+          .mapFromDatabaseResponse(data['${effectivePrefix}sync_state'])),
     );
   }
   factory Config.fromJson(Map<String, dynamic> json,
@@ -28,6 +32,7 @@ class Config extends DataClass implements Insertable<Config> {
     return Config(
       configKey: serializer.fromJson<String>(json['config_key']),
       configValue: serializer.fromJson<String>(json['config_value']),
+      syncState: serializer.fromJson<SyncType>(json['sync_state']),
     );
   }
   factory Config.fromJsonString(String encodedJson,
@@ -40,6 +45,7 @@ class Config extends DataClass implements Insertable<Config> {
     return <String, dynamic>{
       'config_key': serializer.toJson<String>(configKey),
       'config_value': serializer.toJson<String>(configValue),
+      'sync_state': serializer.toJson<SyncType>(syncState),
     };
   }
 
@@ -52,48 +58,62 @@ class Config extends DataClass implements Insertable<Config> {
       configValue: configValue == null && nullToAbsent
           ? const Value.absent()
           : Value(configValue),
+      syncState: syncState == null && nullToAbsent
+          ? const Value.absent()
+          : Value(syncState),
     );
   }
 
-  Config copyWith({String configKey, String configValue}) => Config(
+  Config copyWith({String configKey, String configValue, SyncType syncState}) =>
+      Config(
         configKey: configKey ?? this.configKey,
         configValue: configValue ?? this.configValue,
+        syncState: syncState ?? this.syncState,
       );
   @override
   String toString() {
     return (StringBuffer('Config(')
           ..write('configKey: $configKey, ')
-          ..write('configValue: $configValue')
+          ..write('configValue: $configValue, ')
+          ..write('syncState: $syncState')
           ..write(')'))
         .toString();
   }
 
   @override
-  int get hashCode => $mrjf($mrjc(configKey.hashCode, configValue.hashCode));
+  int get hashCode => $mrjf($mrjc(
+      configKey.hashCode, $mrjc(configValue.hashCode, syncState.hashCode)));
   @override
   bool operator ==(dynamic other) =>
       identical(this, other) ||
       (other is Config &&
           other.configKey == this.configKey &&
-          other.configValue == this.configValue);
+          other.configValue == this.configValue &&
+          other.syncState == this.syncState);
 }
 
 class ConfigCompanion extends UpdateCompanion<Config> {
   final Value<String> configKey;
   final Value<String> configValue;
+  final Value<SyncType> syncState;
   const ConfigCompanion({
     this.configKey = const Value.absent(),
     this.configValue = const Value.absent(),
+    this.syncState = const Value.absent(),
   });
   ConfigCompanion.insert({
     @required String configKey,
     this.configValue = const Value.absent(),
+    this.syncState = const Value.absent(),
   }) : configKey = Value(configKey);
   ConfigCompanion copyWith(
-      {Value<String> configKey, Value<String> configValue}) {
+      {Value<String> configKey,
+      Value<String> configValue,
+      Value<SyncType> syncState}) {
     return ConfigCompanion(
       configKey: configKey ?? this.configKey,
       configValue: configValue ?? this.configValue,
+      syncState: syncState ?? this.syncState,
     );
   }
 }
@@ -120,8 +140,16 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
         $customConstraints: '');
   }
 
+  final VerificationMeta _syncStateMeta = const VerificationMeta('syncState');
+  GeneratedIntColumn _syncState;
+  GeneratedIntColumn get syncState => _syncState ??= _constructSyncState();
+  GeneratedIntColumn _constructSyncState() {
+    return GeneratedIntColumn('sync_state', $tableName, true,
+        $customConstraints: '');
+  }
+
   @override
-  List<GeneratedColumn> get $columns => [configKey, configValue];
+  List<GeneratedColumn> get $columns => [configKey, configValue, syncState];
   @override
   ConfigTable get asDslTable => this;
   @override
@@ -142,6 +170,7 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
       context.handle(_configValueMeta,
           configValue.isAcceptableValue(d.configValue.value, _configValueMeta));
     }
+    context.handle(_syncStateMeta, const VerificationResult.success());
     return context;
   }
 
@@ -162,6 +191,11 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
     if (d.configValue.present) {
       map['config_value'] = Variable<String, StringType>(d.configValue.value);
     }
+    if (d.syncState.present) {
+      final converter = ConfigTable.$converter0;
+      map['sync_state'] =
+          Variable<int, IntType>(converter.mapToSql(d.syncState.value));
+    }
     return map;
   }
 
@@ -170,6 +204,7 @@ class ConfigTable extends Table with TableInfo<ConfigTable, Config> {
     return ConfigTable(_db, alias);
   }
 
+  static TypeConverter<SyncType, int> $converter0 = const SyncTypeConverter();
   @override
   bool get dontWriteConstraints => true;
 }
@@ -1111,6 +1146,7 @@ abstract class _$CustomTablesDb extends GeneratedDatabase {
     return Config(
       configKey: row.readString('config_key'),
       configValue: row.readString('config_value'),
+      syncState: ConfigTable.$converter0.mapToDart(row.readInt('sync_state')),
     );
   }
 
@@ -1190,6 +1226,7 @@ abstract class _$CustomTablesDb extends GeneratedDatabase {
       rowid: row.readInt('rowid'),
       configKey: row.readString('config_key'),
       configValue: row.readString('config_value'),
+      syncState: ConfigTable.$converter0.mapToDart(row.readInt('sync_state')),
     );
   }
 
@@ -1210,7 +1247,7 @@ abstract class _$CustomTablesDb extends GeneratedDatabase {
 
   Future<int> writeConfig(String key, String value) {
     return customInsert(
-      'REPLACE INTO config VALUES (:key, :value)',
+      'REPLACE INTO config (config_key, config_value) VALUES (:key, :value)',
       variables: [Variable.withString(key), Variable.withString(value)],
       updates: {config},
     );
@@ -1224,7 +1261,8 @@ abstract class _$CustomTablesDb extends GeneratedDatabase {
         valueIdx,
         withDefaults,
         myTrigger,
-        OnCreateQuery('INSERT INTO config VALUES (\'key\', \'values\')'),
+        OnCreateQuery(
+            'INSERT INTO config (config_key, config_value) VALUES (\'key\', \'values\')'),
         noIds,
         withConstraints,
         mytable,
@@ -1263,19 +1301,24 @@ class ReadRowIdResult {
   final int rowid;
   final String configKey;
   final String configValue;
+  final SyncType syncState;
   ReadRowIdResult({
     this.rowid,
     this.configKey,
     this.configValue,
+    this.syncState,
   });
   @override
-  int get hashCode => $mrjf(
-      $mrjc(rowid.hashCode, $mrjc(configKey.hashCode, configValue.hashCode)));
+  int get hashCode => $mrjf($mrjc(
+      rowid.hashCode,
+      $mrjc(configKey.hashCode,
+          $mrjc(configValue.hashCode, syncState.hashCode))));
   @override
   bool operator ==(dynamic other) =>
       identical(this, other) ||
       (other is ReadRowIdResult &&
           other.rowid == this.rowid &&
           other.configKey == this.configKey &&
-          other.configValue == this.configValue);
+          other.configValue == this.configValue &&
+          other.syncState == this.syncState);
 }