+++ moor_generator/lib/src/analyzer/moor/create_table_reader.dart
@@ -1,11 +1,13 @@
 import 'package:moor_generator/moor_generator.dart';
 import 'package:moor_generator/src/analyzer/errors.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:moor_generator/src/analyzer/runner/steps.dart';
 import 'package:moor_generator/src/analyzer/sql_queries/type_mapping.dart';
 import 'package:moor_generator/src/model/declarations/declaration.dart';
 import 'package:moor_generator/src/model/used_type_converter.dart';
 import 'package:moor_generator/src/utils/names.dart';
 import 'package:moor_generator/src/utils/string_escaper.dart';
+import 'package:moor_generator/src/utils/type_converter_hint.dart';
 import 'package:recase/recase.dart';
 import 'package:sqlparser/sqlparser.dart';
 
@@ -16,7 +18,7 @@ class CreateTableReader {
 
   CreateTableReader(this.stmt, this.step);
 
-  MoorTable extractTable(TypeMapper mapper) {
+  Future<MoorTable> extractTable(TypeMapper mapper) async {
     Table table;
     try {
       table = SchemaFromCreateTable(moorExtensions: true).read(stmt);
@@ -63,8 +65,9 @@ class CreateTableReader {
         }
 
         if (constraint is MappedBy) {
-          converter = _readTypeConverter(constraint);
-          // don't write MAPPED BY constraints when creating the table
+          converter = await _readTypeConverter(moorType, constraint);
+          // don't write MAPPED BY constraints when creating the table, they're
+          // a convenience feature by the compiler
           continue;
         }
         if (constraint is JsonKey) {
@@ -83,6 +86,10 @@ class CreateTableReader {
       // VIRTUAL TABLE statements - use the entire statement as declaration.
       final declaration =
           MoorColumnDeclaration(column.definition ?? stmt, step.file);
+
+      if (converter != null) {
+        column.applyTypeHint(TypeConverterHint(converter));
+      }
 
       final parsed = MoorColumn(
         type: moorType,
@@ -133,8 +140,21 @@ class CreateTableReader {
     )..parserTable = table;
   }
 
-  UsedTypeConverter _readTypeConverter(MappedBy mapper) {
-    // todo we need to somehow parse the dart expression and check types
-    return null;
+  Future<UsedTypeConverter> _readTypeConverter(
+      ColumnType sqlType, MappedBy mapper) async {
+    final code = mapper.mapper.dartCode;
+    final type = await step.task.backend.resolveTypeOf(step.file.uri, code);
+
+    // todo report lint for any of those cases or when resolveTypeOf throws
+    if (type is! InterfaceType) {
+      return null;
+    }
+
+    final interfaceType = type as InterfaceType;
+    // TypeConverter declares a "D mapToDart(S fromDb);". We need to know D
+    final typeInDart = interfaceType.getMethod('mapToDart').returnType;
+
+    return UsedTypeConverter(
+        expression: code, mappedType: typeInDart, sqlType: sqlType);
   }
 }