+++ moor_generator/lib/src/analyzer/runner/steps/parse_moor.dart
@@ -3,11 +3,8 @@ part of '../steps.dart';
 class ParseMoorStep extends Step {
   final String content;
   final TypeMapper mapper = TypeMapper();
-  /* late final */ InlineDartResolver inlineDartResolver;
 
-  ParseMoorStep(Task task, FoundFile file, this.content) : super(task, file) {
-    inlineDartResolver = InlineDartResolver(this);
-  }
+  ParseMoorStep(Task task, FoundFile file, this.content) : super(task, file);
 
   Future<ParsedMoorFile> parseFile() {
     final parser = MoorParser(this);