+++ moor_generator/lib/src/model/used_type_converter.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:meta/meta.dart';
 import 'package:moor_generator/src/model/table.dart';
@@ -12,10 +11,10 @@ class UsedTypeConverter {
   /// The table using this type converter.
   MoorTable table;
 
-  /// The [Expression] that will construct the type converter at runtime. The
+  /// The expression that will construct the type converter at runtime. The
   /// type converter constructed will map a [mappedType] to the [sqlType] and
   /// vice-versa.
-  final Expression expression;
+  final String expression;
 
   /// The type that will be present at runtime.
   final DartType mappedType;
@@ -23,7 +22,11 @@ class UsedTypeConverter {
   /// The type that will be written to the database.
   final ColumnType sqlType;
 
-  DartType get typeOfConverter => expression.staticType;
+  /// A suitable typename to store an instance of the type converter used here.
+  String get displayNameOfConverter {
+    final sqlDartType = dartTypeNames[sqlType];
+    return 'TypeConverter<${mappedType.getDisplayString()}, $sqlDartType>';
+  }
 
   /// Type converters are stored as static fields in the table that created
   /// them. This will be the field name for this converter.