+++ moor_generator/lib/src/writer/database_writer.dart
@@ -69,7 +69,7 @@ class DatabaseWriter {
 
     // Write fields to access an dao. We use a lazy getter for that.
     for (final dao in db.daos) {
-      final typeName = dao.displayName;
+      final typeName = dao.getDisplayString();
       final getterName = ReCase(typeName).camelCase;
       final databaseImplName = db.fromClass.name;